requirements.txt
praw
discord.py
python-dotenv